split request responsibility from the client
